% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{arch_stat}
\alias{arch_stat}
\title{ARCH LM Statistic}
\usage{
arch_stat(x, lags = 12, demean = TRUE, ...)
}
\arguments{
\item{x}{a univariate time series}

\item{lags}{Number of lags to use in the test}

\item{demean}{Should data have mean removed before test applied?}

\item{...}{Unused.}
}
\value{
A numeric value.
}
\description{
Computes a statistic based on the Lagrange Multiplier (LM) test of Engle (1982) for
autoregressive conditional heteroscedasticity (ARCH). The statistic returned is
the \eqn{R^2}{R^2} value of an autoregressive model of order \code{lags} applied
to \eqn{x^2}{x^2}.
}
