% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{autoplot.tbl_cf}
\alias{autoplot.tbl_cf}
\title{Auto- and Cross- Covariance and -Correlation plots}
\usage{
\method{autoplot}{tbl_cf}(object, level = 95, ...)
}
\arguments{
\item{object}{A tbl_cf object (the result \code{\link[=ACF]{ACF()}}, \code{\link[=PACF]{PACF()}}, or \code{\link[=CCF]{CCF()}}).}

\item{level}{The level of confidence for the blue dashed lines.}

\item{...}{Unused.}
}
\value{
A ggplot object showing the correlations.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{autoplot.tbl_cf()} was soft deprecated in feasts 0.4.2. Please use
\code{ggtime::autoplot.tbl_cf()} instead. Produces a plot of impulse responses
from an impulse response function.

Produces an appropriate plot for the result of  \code{\link[=ACF]{ACF()}}, \code{\link[=PACF]{PACF()}}, or \code{\link[=CCF]{CCF()}}.
}
