% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hctsa_features.R
\name{binarize_mean}
\alias{binarize_mean}
\title{Converts an input vector into a binarized version from software package \code{hctsa}}
\usage{
binarize_mean(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{Unused.}
}
\value{
Time-series values above its mean are given 1, and those below the mean are 0.
}
\description{
Converts an input vector into a binarized version from software package \code{hctsa}
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Yangzhuoran Yang
}
