% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hctsa_features.R
\name{bootstrap_stationarity}
\alias{bootstrap_stationarity}
\title{Bootstrap-based stationarity measure from software package \code{hctsa}}
\usage{
bootstrap_stationarity(x, segment_length = 50, ...)
}
\arguments{
\item{x}{a vector}

\item{segment_length}{the length of local time-series segments to analyse as a positive integer. Can also be a specified character string: "ac2": twice the first zero-crossing of the autocorrelation function}

\item{...}{Unused.}
}
\value{
mean of the first zero-crossings of the autocorrelation function
}
\description{
100 time-series segments of length \code{l} are selected at random from the time series and
the mean of the first zero-crossings of the autocorrelation function in each segment is calculated.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Yangzhuoran Yang
}
