% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical.R
\name{classical_decomposition}
\alias{classical_decomposition}
\title{Classical Seasonal Decomposition by Moving Averages}
\usage{
classical_decomposition(.data, formula, type = c("additive",
  "multiplicative"), ...)
}
\arguments{
\item{.data}{A tsibble.}

\item{formula}{Decomposition specification.}

\item{type}{The type of seasonal component. Can be abbreviated.}

\item{...}{Other arguments passed to \code{\link[stats]{decompose}}.}
}
\description{
Decompose a time series into seasonal, trend and irregular components
  using moving averages.  Deals with additive or multiplicative
  seasonal component.
}
\details{
The additive model used is:
  \deqn{Y_t = T_t + S_t + e_t}{Y[t] = T[t] + S[t] + e[t]}
  The multiplicative model used is:
  \deqn{Y_t = T_t\,S_t\, e_t}{Y[t] = T[t] * S[t] * e[t]}

  The function first determines the trend component using a moving
  average (if \code{filter} is \code{NULL}, a symmetric window with
  equal weights is used), and removes it from the time series.  Then,
  the seasonal figure is computed by averaging, for each time unit, over
  all periods.  The seasonal figure is then centered.   Finally, the error
  component is determined by removing trend and seasonal figure
  (recycled as needed) from the original time series.

  This only works well if \code{x} covers an integer number of complete
  periods.
}
\examples{
USAccDeaths \%>\%
  as_tsibble \%>\%
  classical_decomposition(value ~ season("smallest"))

USAccDeaths \%>\%
  as_tsibble \%>\%
  classical_decomposition(value ~ season(12), type = "mult")

}
