% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{coef_hurst}
\alias{coef_hurst}
\title{Hurst coefficient}
\usage{
coef_hurst(x)
}
\arguments{
\item{x}{a vector. If missing values are present, the largest
contiguous portion of the vector is used.}
}
\value{
A numeric value.
}
\description{
Computes the Hurst coefficient indicating the level of fractional differencing
of a time series.
}
\author{
Rob J Hyndman
}
