% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{crossing_points}
\alias{crossing_points}
\title{Number of crossing points}
\usage{
crossing_points(x, ...)
}
\arguments{
\item{x}{a univariate time series}

\item{...}{Unused.}
}
\value{
A numeric value.
}
\description{
Computes the number of times a time series crosses the median.
}
