% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{entropy}
\alias{entropy}
\title{Spectral entropy of a time series}
\usage{
entropy(x, ...)
}
\arguments{
\item{x}{a vector}
}
\value{
A numeric value.
}
\description{
Computes the spectral entropy of a time series
}
\author{
Rob J Hyndman
}
