% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_intermittent}
\alias{feat_intermittent}
\title{Intermittency features}
\usage{
feat_intermittent(x)
}
\arguments{
\item{x}{A vector to extract features from.}
}
\value{
A vector of named features:
\itemize{
\item zero_run_mean: The average interval between non-zero observations
\item nonzero_squared_cv: The squared coefficient of variation of non-zero observations
\item zero_start_prop: The proportion of data which starts with zero
\item zero_end_prop: The proportion of data which ends with zero
}
}
\description{
Computes various measures that can indicate the presence and structures of
intermittent data.
}
\references{
Kostenko, A. V., & Hyndman, R. J. (2006). A note on the categorization of
demand patterns. \emph{Journal of the Operational Research Society}, 57(10),
1256-1257.
}
