% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{features}
\alias{features}
\title{Extract features from a dataset}
\usage{
features(.data, ...)
}
\arguments{
\item{.data}{A dataset (tsibble)}

\item{...}{Additional arguments to be passed to measures that use it.}
}
\description{
Extract features from a dataset
}
