% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hctsa_features.R
\docType{data}
\name{features_compengine}
\alias{features_compengine}
\title{CompEngine feature set}
\format{An object of class \code{list} of length 5.}
\usage{
features_compengine
}
\arguments{
\item{x}{a vector}
}
\value{
a vector with CompEngine features
}
\description{
Calculate the features that have been used in CompEngine database, using method introduced in package
\code{hctsa}.
}
\details{
The features involved can be grouped as \code{autocorrelation},
\code{prediction}, \code{stationarity}, \code{distribution}, and \code{scaling}.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\seealso{
\code{\link{features_autocorrelation}}

\code{\link{features_prediction}}

\code{\link{features_stationarity}}

\code{\link{features_distribution}}

\code{\link{features_scaling}}
}
\author{
Yangzhuoran Yang
}
\keyword{datasets}
