% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{flat_spots}
\alias{flat_spots}
\title{Number of flat spots}
\usage{
flat_spots(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A numeric value.
}
\description{
Number of flat spots in a time series
}
\author{
Earo Wang and Rob J Hyndman
}
