% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hctsa_features.R
\name{fluctuation_analysis}
\alias{fluctuation_analysis}
\title{Implements fluctuation analysis from software package \code{hctsa}}
\usage{
fluctuation_analysis(x)
}
\description{
Fits a polynomial of order 1 and then returns the
range. The order of fluctuations is 2, corresponding to root mean
square fluctuations.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Yangzhuoran Yang
}
