% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{gg_irf}
\alias{gg_irf}
\title{Plot impulse response functions}
\usage{
gg_irf(data, y = all_of(measured_vars(data)))
}
\arguments{
\item{data}{A tsibble with impulse responses}

\item{y}{The impulse response variables to plot (defaults to all measured variables).}
}
\value{
A ggplot object of the impulse responses.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{gg_irf()} was soft deprecated in feasts 0.4.2. Please use \code{ggtime::gg_irf()} instead.
Produces a plot of impulse responses from an impulse response function.
}
