% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{gglagplot}
\alias{gglagplot}
\alias{gglagplot.tbl_ts}
\title{Time series lag ggplots}
\usage{
gglagplot(x, ...)

\method{gglagplot}{tbl_ts}(x, var = NULL, period = "smallest",
  lags = 1:16, ...)
}
\arguments{
\item{x}{a time series object (type \code{ts}).}

\item{...}{Not used (for consistency with lag.plot)}

\item{var}{The variable to plot (a bare expression). If NULL, it will
automatically selected from the data.}

\item{period}{The seasonal period to display.}

\item{lags}{A vector of lags to display as facets.}
}
\value{
None.
}
\description{
Plots a lag plot using ggplot.
}
\details{
\dQuote{gglagplot} will plot time series against lagged versions of
themselves. Helps visualising 'auto-dependence' even when auto-correlations
vanish.

\dQuote{gglagchull} will layer convex hulls of the lags, layered on a single
plot. This helps visualise the change in 'auto-dependence' as lags increase.
}
\examples{
library(tsibble)
tsibbledata::ausretail \%>\%
  filter(
    State == "Victoria",
    Industry == "Cafes, restaurants and catering services"
  ) \%>\%
  gglagplot(Turnover)

}
\seealso{
\code{\link[stats]{lag.plot}}
}
