% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ggseasonplot}
\alias{ggseasonplot}
\alias{ggseasonplot.tbl_ts}
\title{Seasonal plot}
\usage{
ggseasonplot(x, ...)

\method{ggseasonplot}{tbl_ts}(x, var = NULL, period = "largest",
  facet_period = NULL, ...)
}
\arguments{
\item{x}{a numeric vector or time series of class \code{ts}.}

\item{...}{additional arguments to \code{\link[graphics]{plot}}.}
}
\value{
None.
}
\description{
Plots a seasonal plot as described in Hyndman and Athanasopoulos (2014,
chapter 2). This is like a time plot except that the data are plotted
against the seasons in separate years.
}
\examples{
tsibbledata::ausretail \%>\%
  filter(
    State == "Victoria",
    Industry == "Cafes, restaurants and catering services"
  ) \%>\%
  ggseasonplot(Turnover)

}
\references{
Hyndman and Athanasopoulos (2018) \emph{Forecasting: principles
and practice}, 2nd edition, OTexts: Melbourne, Australia.
\url{https://OTexts.org/fpp2/}
}
\seealso{
\code{\link[stats]{monthplot}}
}
