% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ggsubseriesplot}
\alias{ggsubseriesplot}
\alias{ggsubseriesplot.tbl_ts}
\title{Create a seasonal subseries ggplot}
\usage{
ggsubseriesplot(x, ...)

\method{ggsubseriesplot}{tbl_ts}(x, var = NULL, period = NULL, ...)
}
\arguments{
\item{x}{a time series object (type \code{ts}).}

\item{...}{Not used (for consistency with monthplot)}

\item{var}{The variable to plot (a bare expression). If NULL, it will
automatically selected from the data.}

\item{period}{The seasonal period to display.}
}
\value{
Returns an object of class \code{ggplot}.
}
\description{
Plots a subseries plot using ggplot. Each season is plotted as a separate
mini time series. The blue lines represent the mean of the observations
within each season.
}
\details{
The \code{ggmonthplot} function is simply a wrapper for
\code{ggsubseriesplot} as a convenience for users familiar with
\code{\link[stats]{monthplot}}.
}
\examples{
library(tsibble)
tsibbledata::ausretail \%>\%
  filter(
    State == "Victoria",
    Industry == "Cafes, restaurants and catering services"
  ) \%>\%
  ggsubseriesplot(Turnover)

}
\seealso{
\code{\link[stats]{monthplot}}
}
