% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ggtsdisplay}
\alias{ggtsdisplay}
\alias{ggtsdisplay.tbl_ts}
\title{Time series display}
\usage{
ggtsdisplay(x, ...)

\method{ggtsdisplay}{tbl_ts}(x, var = NULL, plot_type = c("partial",
  "histogram", "scatter", "spectrum"), lag_max = NULL, ...)
}
\arguments{
\item{x}{a numeric vector or time series of class \code{ts}.}

\item{...}{additional arguments to \code{\link[stats]{acf}}.}

\item{var}{The variable to plot (a bare expression). If NULL, it will
automatically selected from the data.}

\item{lags}{A vector of lags to display as facets.}
}
\value{
None.
}
\description{
Plots a time series along with its acf and either its pacf, lagged
scatterplot or spectrum.
}
\details{
\code{ggtsdisplay} will produce the equivalent plot using ggplot graphics.
}
\examples{
library(tsibble)
tsibbledata::ausretail \%>\%
  filter(
    State == "Victoria",
    Industry == "Cafes, restaurants and catering services"
  ) \%>\%
  ggtsdisplay(Turnover)

}
\references{
Hyndman and Athanasopoulos (2018) \emph{Forecasting: principles
and practice}, 2nd edition, OTexts: Melbourne, Australia.
\url{https://OTexts.org/fpp2/}
}
\seealso{
\code{\link[stats]{plot.ts}}, \code{\link{Acf}},
\code{\link[stats]{spec.ar}}
}
