% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hctsa_features.R
\name{histogram_mode}
\alias{histogram_mode}
\title{Mode of a data vector from software package \code{hctsa}}
\usage{
histogram_mode(x, bins = 10, ...)
}
\arguments{
\item{x}{a vector}

\item{bins}{the number of bins to use.}

\item{...}{Unused.}
}
\value{
the mode
}
\description{
Measures the mode of the data vector binned by the specified number of bins.
The value calculated is different from \code{hctsa} and \code{CompEngine} as the histogram edges are calculated differently.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
