% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hctsa_features.R
\name{localsimple_taures}
\alias{localsimple_taures}
\title{The first zero crossing of the autocorrelation function of the residuals from simple local time-series forecasting from software package \code{hctsa}}
\usage{
localsimple_taures(x, fc_method = c("mean", "lfit"),
  train_length = NULL)
}
\arguments{
\item{x}{a vector}

\item{fc_method}{the forecasting method, default to \code{mean}.
\code{mean}: local mean prediction using the past \code{train_length} values.
\code{lfit}: local linear prediction using the past \code{train_length} values.}

\item{train_length}{the number of values to use to forecast the next value.
Default to 1 when using method \code{mean} and the first zero auto-correlation
(via \code{\link[=firstzero_ac]{firstzero_ac()}} when using method \code{lfit}.}
}
\value{
The first zero crossing of the autocorrelation function of the residuals
}
\description{
Simple predictors using the past \code{train_length} values of the time series to
predict its next value.
}
