% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{longest_flat_spot}
\alias{longest_flat_spot}
\alias{n_flat_spots}
\title{Longest flat spot length}
\usage{
longest_flat_spot(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A numeric value.
}
\description{
"Flat spots” are computed by dividing the sample space of a time series into
ten equal-sized intervals, and computing the maximum run length within any
single interval.
}
\author{
Earo Wang and Rob J Hyndman
}
