% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hctsa_features.R
\name{motiftwo_entro3}
\alias{motiftwo_entro3}
\title{Local motifs in a binary symbolization of the time series from software package \code{hctsa}}
\usage{
motiftwo_entro3(x)
}
\arguments{
\item{x}{a vector}
}
\value{
Entropy of words in the binary alphabet of length 3.
}
\description{
Coarse-graining is performed. Time-series values above its mean are given 1,
and those below the mean are 0.
}
\examples{
motiftwo_entro3(WWWusage)

}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Yangzhuoran Yang
}
