% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{pacf_features}
\alias{pacf_features}
\title{Partial autocorrelation-based features}
\usage{
pacf_features(x, .period = 1, ...)
}
\arguments{
\item{x}{a univariate time series}

\item{.period}{The seasonal period (optional)}

\item{...}{Unused.}
}
\value{
A vector of 3 values: Sum of squared of first 5
partial autocorrelation coefficients of the original series, first differenced
series and twice-differenced series.
For seasonal data, the partial autocorrelation coefficient at the first seasonal
lag is also returned.
}
\description{
Computes various measures based on partial autocorrelation coefficients of the
original series, first-differenced series and second-differenced series.
}
\author{
Thiyanga Talagala
}
