% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hctsa_features.R
\name{pred_features}
\alias{pred_features}
\title{The prediction feature set from software package \code{hctsa}}
\usage{
pred_features(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector with autocorrelation features
}
\description{
Calculate the features that grouped as prediction set,
which have been used in CompEngine database, using method introduced in package \code{hctsa}.
}
\details{
Features in this set are \code{localsimple_mean_ac},
\code{localsimple_lfit_ac},
and \code{sampen_first}.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\seealso{
\code{\link{localsimple_taures}}

\code{\link{sampen_first}}
}
\author{
Yangzhuoran Yang
}
