% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gg_season}
\alias{gg_subseries}
\alias{gg_lag}
\alias{gg_tsdisplay}
\alias{gg_tsresiduals}
\alias{gg_arma}
\alias{gg_irf}
\alias{\%>\%}
\alias{as_tsibble}
\alias{autoplot}
\alias{autolayer}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{ggplot2}{\code{\link[ggplot2]{autolayer}}, \code{\link[ggplot2]{autoplot}}}

  \item{ggtime}{\code{\link[ggtime]{gg_arma}}, \code{\link[ggtime]{gg_irf}}, \code{\link[ggtime]{gg_lag}}, \code{\link[ggtime]{gg_season}}, \code{\link[ggtime]{gg_subseries}}, \code{\link[ggtime]{gg_tsdisplay}}, \code{\link[ggtime]{gg_tsresiduals}}}

  \item{tsibble}{\code{\link[tsibble:as-tsibble]{as_tsibble}}}
}}

