% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{scale_cf_lag}
\alias{scale_cf_lag}
\alias{scale_x_cf_lag}
\title{lagged datetime scales
This set of scales defines new scales for lagged time structures.}
\usage{
scale_x_cf_lag(...)
}
\arguments{
\item{...}{Further arguments to be passed on to scale_x_continuous()}
}
\value{
A ggproto object inheriting from \code{Scale}
}
\description{
lagged datetime scales
This set of scales defines new scales for lagged time structures.
}
\keyword{internal}
