% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hctsa_features.R
\name{stationarity_features}
\alias{stationarity_features}
\title{The stationarity feature set from software package \code{hctsa}}
\usage{
stationarity_features(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector with autocorrelation features
}
\description{
Calculate the features that grouped as stationarity set,
which have been used in CompEngine database, using method introduced in package \code{hctsa}.
}
\details{
Features in this set are \code{sd_deriv_1},
\code{spreadrandomlocal_meantaul_50},
and \code{spreadrandomlocal_meantaul_ac2}.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\seealso{
\code{\link{sd_deriv_1}}

\code{\link{bootstrap_stationarity}}
}
\author{
Yangzhuoran Yang
}
