% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{stl_features}
\alias{stl_features}
\title{STL features}
\usage{
stl_features(x, .period, s.window = 13, ...)
}
\arguments{
\item{x}{A vector to extract features from.}

\item{.period}{The period of the seasonality.}

\item{s.window}{The seasonal window of the data (passed to \code{\link[stats:stl]{stats::stl()}})}

\item{...}{Further arguments passed to \code{\link[stats:stl]{stats::stl()}}}
}
\description{
Computes a variety of measures extracted from an STL decomposition of the
time series. This includes details about the strength of trend and seasonality.
}
\seealso{
\link{https://otexts.com/fpp3/seasonal-strength.html}
}
