% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition_str.R
\name{STR}
\alias{STR}
\title{Seasonal-Trend Decomposition Using Regression}
\usage{
STR(formula, iterations = 2, robust = FALSE, ...)
}
\arguments{
\item{formula}{Decomposition specification (see "Specials" section).}

\item{iterations}{Number of iterations to use to refine the seasonal component.}

\item{...}{Not used.}
}
\value{
A \code{\link[fabletools:dable]{fabletools::dable()}} containing the decomposed trend, seasonality
and remainder from the STL decomposition.
}
\description{
Decompose a time series with a regular time index using regression. This
method allows use of multiple seasonal and cyclic components, trends, and
covariates. It supports seasonalities with non-integer periods and complex
topologies.
}
\section{Specials}{


\subsection{trend}{
The \code{trend} special is used to specify the trend extraction parameters.
\preformatted{
trend(window, degree, jump)
}

\tabular{ll}{
\code{window} \tab The span (in lags) of the loess window, which should be odd. If NULL, the default, nextodd(ceiling((1.5*period) / (1-(1.5/s.window)))), is taken.\cr
\code{degree} \tab The degree of locally-fitted polynomial. Should be zero or one. \cr
\code{jump}   \tab Integers at least one to increase speed of the respective smoother. Linear interpolation happens between every \code{jump}th value.
}
}

\subsection{season}{
The \code{season} special is used to specify the season extraction parameters.
\preformatted{
season(period = NULL, window = NULL, degree, jump)
}

\tabular{ll}{
\code{period} \tab The periodic nature of the seasonality. This can be either a number indicating the number of observations in each seasonal period, or text to indicate the duration of the seasonal window (for example, annual seasonality would be "1 year").\cr
\code{window} \tab The span (in lags) of the loess window, which should be odd. If the \code{window} is set to \code{"periodic"} or \code{Inf}, the seasonal pattern will be fixed. The window size should be odd and at least 7, according to Cleveland et al. The default (NULL) will choose an appropriate default, for a dataset with one seasonal pattern this would be 11, the second larger seasonal window would be 15, then 19, 23, ... onwards.\cr
\code{degree} \tab The degree of locally-fitted polynomial. Should be zero or one. \cr
\code{jump}   \tab Integers at least one to increase speed of the respective smoother. Linear interpolation happens between every \code{jump}th value.
}
}

\subsection{lowpass}{
The \code{lowpass} special is used to specify the low-pass filter parameters.
\preformatted{
lowpass(window, degree, jump)
}

\tabular{ll}{
\code{window} \tab The span (in lags) of the loess window of the low-pass filter used for each subseries. Defaults to the smallest odd integer greater than or equal to the seasonal \code{period} which is recommended since it prevents competition between the trend and seasonal components. If not an odd integer its given value is increased to the next odd one. \cr
\code{degree} \tab The degree of locally-fitted polynomial. Must be zero or one. \cr
\code{jump}   \tab Integers at least one to increase speed of the respective smoother. Linear interpolation happens between every \code{jump}th value.
}
}
}

\examples{
as_tsibble(USAccDeaths) \%>\%
  model(STR(value ~ trend(window = 10))) \%>\%
  components()

}
\references{
Alexander Dokumentov, Rob J. Hyndman (2022) STR: Seasonal-Trend Decomposition Using Regression. INFORMS Journal on Data Science 1(1):50-62. https://doi.org/10.1287/ijds.2021.0004
}
