% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{lumpiness}
\alias{lumpiness}
\alias{stability}
\title{Time series features based on tiled windows}
\usage{
lumpiness(x, .size = NULL, .period = 1, ...)

stability(x, .size = NULL, .period = 1, ...)
}
\arguments{
\item{x}{a univariate time series}

\item{.size}{size of sliding window, if NULL \code{.size} will be automatically chosen using \code{.period}}

\item{.period}{The seasonal period (optional)}

\item{...}{Unused.}
}
\value{
A numeric vector of length 2 containing a measure of lumpiness and
a measure of stability.
}
\description{
Computes feature of a time series based on tiled (non-overlapping) windows.
Means or variances are produced for all tiled windows. Then stability is
the variance of the means, while lumpiness is the variance of the variances.
}
\author{
Earo Wang and Rob J Hyndman
}
