% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hctsa_features.R
\name{trev_num}
\alias{trev_num}
\title{Normalized nonlinear autocorrelation, the numerator of the trev function of a time series from software package \code{hctsa}}
\usage{
trev_num(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the numerator of the trev function of a time series
}
\description{
Calculates the numerator of the trev function, a normalized nonlinear autocorrelation,
The time lag is set to 1.
}
\examples{
trev_num(WWWusage)

}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Yangzhuoran Yang
}
