% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibblestats.R
\docType{package}
\name{tsibblestats-package}
\alias{tsibblestats}
\alias{tsibblestats-package}
\title{tsibblestats: Analysis of tidy time series}
\description{
A collection of tsibble analysis tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverts/tsibblestats}
  \item Report bugs at \url{https://github.com/tidyverts/tsibblestats/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} [copyright holder]

Authors:
\itemize{
  \item Mitchell O'Hara-Wild
  \item Earo Wang
}

}
\keyword{package}
