% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{unitroot_kpss}
\alias{unitroot_kpss}
\alias{unitroot_pp}
\title{Unit root tests}
\usage{
unitroot_kpss(x, ...)

unitroot_pp(x, ...)
}
\arguments{
\item{x}{A vector to be tested for the unit root.}

\item{...}{Unused.}
}
\description{
Performs a test for the existence of a unit root in the vector.
}
\details{
\code{unitroot_kpss} computes the statistic for the Kwiatkowski et al. unit root test with linear trend and lag 1.
\code{unitroot_pp} computes the statistic for the `'Z-tau'' version of Phillips & Perron unit root test with constant trend and lag 1.
}
