% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x11.R
\name{X11}
\alias{X11}
\title{X11 seasonal decomposition}
\usage{
X11(.data, formula, type = c("additive", "multiplicative"), ...)
}
\arguments{
\item{.data}{A tsibble.}

\item{formula}{Decomposition specification.}

\item{type}{The type of decomposition: additive or multiplicative}

\item{...}{Other arguments passed to \code{\link[seasonal:seas]{seasonal::seas()}}.}
}
\description{
Applies a seasonal adjustment by an enhanced version of the methodology of the
Census Bureau X-11 and X-11Q programs. The type of seasonal decomposition
(additive or multiplicative) can be controlled using the \code{type} argument.
}
\section{Specials}{


\subsection{season}{
The \code{season} special is used to specify seasonal attributes of the decomposition.
\preformatted{
season(period = NULL)
}

\tabular{ll}{
\code{period}   \tab The periodic nature of the seasonality. X11 decompositions only support monthly (\code{period = 12}) and quarterly (\code{period = 4}) seasonal patterns.
}
}
}

\examples{
tsibbledata::aus_production \%>\% X11(Beer)

}
\references{
Dagum, E. B., & Bianconcini, S. (2016) "Seasonal adjustment methods and real
time trend-cycle estimation". \emph{Springer}.

X11 Documentation from the seasonal package's website:
http://www.seasonal.website/seasonal.html#input

Official X-13ARIMA-SEATS manual: https://www.census.gov/ts/x13as/docX13ASHTML.pdf
}
\seealso{
\code{\link[seasonal:seas]{seasonal::seas()}}
}
