% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{ACF}
\alias{ACF}
\alias{PACF}
\alias{CCF}
\title{(Partial) Autocorrelation and Cross-Correlation Function Estimation}
\usage{
ACF(.data, value = NULL, ...)

PACF(.data, value = NULL, ...)

CCF(.data, value1, value2, ...)
}
\arguments{
\item{.data}{A tsibble}

\item{value, value1, value2}{The column(s) from the tsibble used to compute the ACF, PACF or CCF.}

\item{...}{Further arguments to be passed on to acf, pacf, and ccf}
}
\value{
The \code{ACF}, \code{PACF} and \code{CCF} functions return objects
of class "tbl_cf", which is a tibble containing the correlations computed.
}
\description{
The function \code{ACF} computes an estimate of the autocorrelation function
of a (possibly multivariate) tsibble. Function \code{PACF} computes an estimate
of the partial autocorrelation function of a (possibly multivariate) tsibble.
Function \code{CCF} computes the cross-correlation or cross-covariance of two columns
from a tsibble.
}
\details{
The functions improve the \code{\link[stats]{acf}},
\code{\link[stats]{pacf}} and \code{\link[stats]{ccf}} functions. The main
differences are that \code{ACF} does not plot a spike at lag 0 when
\code{type=="correlation"} (which is redundant) and the horizontal axes show
lags in time units rather than seasonal units.

The tapered versions implement the ACF and PACF estimates and plots
described in Hyndman (2015), based on the banded and tapered estimates of
autocovariance proposed by McMurry and Politis (2010).
}
\examples{
tsibbledata::elecdemand \%>\% ACF(Temperature)

tsibbledata::elecdemand \%>\% PACF(Demand)
tsibbledata::UKLungDeaths \%>\% CCF(mdeaths, fdeaths)
}
\references{
Hyndman, R.J. (2015). Discussion of ``High-dimensional
autocovariance matrices and optimal linear prediction''. \emph{Electronic
Journal of Statistics}, 9, 792-796.

McMurry, T. L., & Politis, D. N. (2010). Banded and tapered estimates for
autocovariance matrices and the linear process bootstrap. \emph{Journal of
Time Series Analysis}, 31(6), 471-482.
}
\seealso{
\code{\link[stats]{acf}}, \code{\link[stats]{pacf}},
\code{\link[stats]{ccf}}
}
\author{
Mitchell O'Hara-Wild and Rob J Hyndman
}
