% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{ACF}
\alias{ACF}
\alias{PACF}
\alias{CCF}
\title{(Partial) Autocorrelation and Cross-Correlation Function Estimation}
\usage{
ACF(
  .data,
  y,
  ...,
  lag_max = NULL,
  type = c("correlation", "covariance", "partial"),
  na.action = na.contiguous,
  demean = TRUE,
  tapered = FALSE
)

PACF(.data, y, ..., lag_max = NULL, na.action = na.contiguous, tapered = FALSE)

CCF(
  .data,
  y,
  x,
  ...,
  lag_max = NULL,
  type = c("correlation", "covariance"),
  na.action = na.contiguous
)
}
\arguments{
\item{.data}{A tsibble}

\item{...}{The column(s) from the tsibble used to compute the ACF, PACF or CCF.}

\item{lag_max}{maximum lag at which to calculate the acf. Default is 10*log10(N/m)
where N is the number of observations and m the number of series. Will be
automatically limited to one less than the number of observations in the series.}

\item{type}{character string giving the type of \abbr{acf} to be computed.
    Allowed values are
    \code{"correlation"} (the default), \code{"covariance"} or
    \code{"partial"}.  Will be partially matched.}

\item{na.action}{function to be called to handle missing
    values. \code{na.pass} can be used.}

\item{demean}{logical.  Should the covariances be about the sample
    means?}

\item{tapered}{Produces banded and tapered estimates of the (partial) autocorrelation.}

\item{x, y}{a univariate or multivariate (not \code{ccf}) numeric time
    series object or a numeric vector or matrix, or an \code{"acf"} object.}
}
\value{
The \code{ACF}, \code{PACF} and \code{CCF} functions return objects
of class "tbl_cf", which is a tsibble containing the correlations computed.
}
\description{
The function \code{ACF} computes an estimate of the autocorrelation function
of a (possibly multivariate) tsibble. Function \code{PACF} computes an estimate
of the partial autocorrelation function of a (possibly multivariate) tsibble.
Function \code{CCF} computes the cross-correlation or cross-covariance of two columns
from a tsibble.
}
\details{
The functions improve the \code{\link[stats:acf]{stats::acf()}}, \code{\link[stats:acf]{stats::pacf()}} and
\code{\link[stats:acf]{stats::ccf()}} functions. The main differences are that \code{ACF} does not plot
the exact correlation at lag 0 when \code{type=="correlation"} and
the horizontal axes show lags in time units rather than seasonal units.

The resulting tables from these functions can also be plotted using
\code{\link[=autoplot.tbl_cf]{autoplot.tbl_cf()}}.
}
\examples{
library(tsibble)
library(tsibbledata)
library(dplyr)

vic_elec \%>\% ACF(Temperature)

vic_elec \%>\% ACF(Temperature) \%>\% autoplot()

vic_elec \%>\% PACF(Temperature)

vic_elec \%>\% PACF(Temperature) \%>\% autoplot()

global_economy \%>\%
  filter(Country == "Australia") \%>\%
  CCF(GDP, Population)

global_economy \%>\%
  filter(Country == "Australia") \%>\%
  CCF(GDP, Population) \%>\%
  autoplot()

}
\references{
Hyndman, R.J. (2015). Discussion of "High-dimensional
autocovariance matrices and optimal linear prediction". \emph{Electronic
Journal of Statistics}, 9, 792-796.

McMurry, T. L., & Politis, D. N. (2010). Banded and tapered estimates for
autocovariance matrices and the linear process bootstrap. \emph{Journal of
Time Series Analysis}, 31(6), 471-482.
}
\seealso{
\code{\link[stats:acf]{stats::acf()}}, \code{\link[stats:acf]{stats::pacf()}}, \code{\link[stats:acf]{stats::ccf()}}
}
\author{
Mitchell O'Hara-Wild and Rob J Hyndman
}
