% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasts.R
\docType{package}
\name{feasts-package}
\alias{feasts}
\alias{feasts-package}
\title{feasts: Feature Extraction And Statistics for Time Series}
\description{
A collection of tsibble analysis tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverts/feasts}
  \item Report bugs at \url{https://github.com/tidyverts/feasts/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} [copyright holder]

Authors:
\itemize{
  \item Mitchell O'Hara-Wild
  \item Earo Wang
}

}
\keyword{package}
