% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasts.R
\docType{package}
\name{feasts-package}
\alias{feasts}
\alias{feasts-package}
\title{feasts: Feature Extraction And Statistics for Time Series}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides a collection of features, decompositions, statistical summaries and graphics functions for the analysing tidy time series data. The package name 'feasts' is an acryonym comprising of its key features: Feature Extraction And Statistics for Time Series.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://feasts.tidyverts.org/}
  \item Report bugs at \url{https://github.com/tidyverts/feasts/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Rob Hyndman
  \item Earo Wang
}

Other contributors:
\itemize{
  \item Di Cook [contributor]
}

}
\keyword{package}
