% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_spectral}
\alias{feat_spectral}
\title{Spectral features of a time series}
\usage{
feat_spectral(x, .period = 1, ...)
}
\arguments{
\item{x}{a univariate time series}

\item{.period}{The seasonal period.}

\item{...}{Further arguments for \code{\link[ForeCA:spectral_entropy]{ForeCA::spectral_entropy()}}}
}
\value{
A numeric value.
}
\description{
Computes the spectral entropy of a time series
}
\author{
Rob J Hyndman
}
