% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{features}
\alias{features}
\alias{features.tbl_ts}
\title{Extract features from a dataset}
\usage{
features(.tbl, ...)

\method{features}{tbl_ts}(.tbl, .vars = NULL, features = list(), ...)
}
\arguments{
\item{.tbl}{A dataset (tsibble/tibble)}

\item{...}{Additional arguments to be passed to measures.}

\item{.vars}{A list of columns generated by \code{\link[=vars]{vars()}},
a character vector of column names, a numeric vector of column
positions, or \code{NULL}.}

\item{features}{A list of functions (or lambda expressions) for the features to compute.}
}
\description{
Extract features from a dataset
}
