% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{features_spectral}
\alias{features_spectral}
\title{Spectral features of a time series}
\usage{
features_spectral(x, ...)
}
\arguments{
\item{x}{a univariate time series}
}
\value{
A numeric value.
}
\description{
Computes the spectral entropy of a time series
}
\author{
Rob J Hyndman
}
