% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{gg_irf}
\alias{gg_irf}
\title{Plot impulse response functions}
\usage{
gg_irf(data, y = all_of(measured_vars(data)))
}
\arguments{
\item{data}{A tsibble with impulse responses}
}
\value{
A ggplot object of the impulse responses.
}
\description{
Produces a plot of impulse responses from an impulse response function.
}
