% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{gg_lag}
\alias{gg_lag}
\title{Time series lag ggplots}
\usage{
gg_lag(data, y = NULL, period = NULL, lags = 1:9, geom = c("path",
  "point"), ...)
}
\arguments{
\item{data}{A tidy time series object (tsibble)}

\item{y}{The variable to plot (a bare expression). If NULL, it will
automatically selected from the data.}

\item{period}{The seasonal period to display.}

\item{lags}{A vector of lags to display as facets.}

\item{geom}{The geometry used to display the data.}

\item{...}{Not used (for consistency with lag.plot)}
}
\value{
None.
}
\description{
Plots a lag plot using ggplot.
}
\details{
\dQuote{gglagplot} will plot time series against lagged versions of
themselves. Helps visualising 'auto-dependence' even when auto-correlations
vanish.

\dQuote{gglagchull} will layer convex hulls of the lags, layered on a single
plot. This helps visualise the change in 'auto-dependence' as lags increase.
}
\examples{
library(tsibble)
tsibbledata::aus_retail \%>\%
  filter(
    State == "Victoria",
    Industry == "Cafes, restaurants and catering services"
  ) \%>\%
  gg_lag(Turnover)

}
\seealso{
\code{\link[stats]{lag.plot}}
}
