% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{gg_season}
\alias{gg_season}
\title{Seasonal plot}
\usage{
gg_season(data, y = NULL, period = NULL, facet_period, max_col = 15,
  polar = FALSE, labels = c("none", "left", "right", "both"))
}
\arguments{
\item{data}{A tidy time series object (tsibble)}

\item{y}{The variable to plot (a bare expression). If NULL, it will
automatically selected from the data.}

\item{period}{The seasonal period to display.}

\item{facet_period}{A secondary seasonal period to facet by
(typically smaller than period).}

\item{max_col}{The maximum number of colours to display on the plot. If the
number of seasonal periods in the data is larger than \code{max_col}, the plot
will not include a colour. Use \code{max_col = 0} to never colour the lines, or Inf
to always colour the lines.}

\item{polar}{If TRUE, the season plot will be shown on polar coordinates.}

\item{labels}{Position of the labels for seasonal period identifier.}
}
\value{
None.
}
\description{
Plots a seasonal plot as described in Hyndman and Athanasopoulos (2014,
chapter 2). This is like a time plot except that the data are plotted
against the seasons in separate years.
}
\examples{
library(tsibble)
tsibbledata::aus_retail \%>\%
  filter(
    State == "Victoria",
    Industry == "Cafes, restaurants and catering services"
  ) \%>\%
  gg_season(Turnover)

}
\references{
Hyndman and Athanasopoulos (2018) \emph{Forecasting: principles
and practice}, 2nd edition, OTexts: Melbourne, Australia.
\url{https://OTexts.org/fpp2/}
}
\seealso{
\code{\link[stats]{monthplot}}
}
