% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{gg_subseries}
\alias{gg_subseries}
\title{Create a seasonal subseries ggplot}
\usage{
gg_subseries(data, y = NULL, period = NULL)
}
\arguments{
\item{data}{A tidy time series object (tsibble)}

\item{y}{The variable to plot (a bare expression). If NULL, it will
automatically selected from the data.}

\item{period}{The seasonal period to display.}
}
\value{
Returns an object of class \code{ggplot}.
}
\description{
Plots a subseries plot using ggplot. Each season is plotted as a separate
mini time series. The blue lines represent the mean of the observations
within each season.
}
\details{
The \code{ggmonthplot} function is simply a wrapper for
\code{ggsubseriesplot} as a convenience for users familiar with
\code{\link[stats]{monthplot}}.
}
\examples{
library(tsibble)
tsibbledata::aus_retail \%>\%
  filter(
    State == "Victoria",
    Industry == "Cafes, restaurants and catering services"
  ) \%>\%
  gg_subseries(Turnover)

}
\seealso{
\code{\link[stats]{monthplot}}
}
