% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ggseasonplot}
\alias{ggseasonplot}
\alias{ggseasonplot.tbl_ts}
\title{Seasonal plot}
\usage{
ggseasonplot(x, ...)

\method{ggseasonplot}{tbl_ts}(x, var = NULL, period = NULL,
  facet_period, polar = FALSE, labels = c("none", "left", "right",
  "both"), ...)
}
\arguments{
\item{x}{A time series object}

\item{...}{Additional arguments passed to methods}

\item{var}{The variable to plot (a bare expression). If NULL, it will
automatically selected from the data.}

\item{period}{The seasonal period to display.}

\item{facet_period}{A secondary seasonal period to facet by
(typically smaller than period).}

\item{polar}{If TRUE, the seasonplot will be shown on polar coordinates.}

\item{labels}{Position of the labels for seasonal period identifier.}
}
\value{
None.
}
\description{
Plots a seasonal plot as described in Hyndman and Athanasopoulos (2014,
chapter 2). This is like a time plot except that the data are plotted
against the seasons in separate years.
}
\examples{
library(tsibble)
tsibbledata::aus_retail \%>\%
  filter(
    State == "Victoria",
    Industry == "Cafes, restaurants and catering services"
  ) \%>\%
  ggseasonplot(Turnover)

}
\references{
Hyndman and Athanasopoulos (2018) \emph{Forecasting: principles
and practice}, 2nd edition, OTexts: Melbourne, Australia.
\url{https://OTexts.org/fpp2/}
}
\seealso{
\code{\link[stats]{monthplot}}
}
