% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ggtsdisplay}
\alias{ggtsdisplay}
\alias{ggtsdisplay.tbl_ts}
\title{Ensemble of time series displays}
\usage{
ggtsdisplay(x, ...)

\method{ggtsdisplay}{tbl_ts}(x, var = NULL, plot_type = c("partial",
  "histogram", "scatter", "spectrum"), lag_max = NULL, ...)
}
\arguments{
\item{x}{A time series object}

\item{...}{Arguments for subsequent plotting methods.}

\item{var}{The variable to plot (a bare expression). If NULL, it will
automatically selected from the data.}

\item{plot_type}{type of plot to include in lower right corner.}

\item{lag_max}{maximum lag at which to calculate the acf. Default is 10*log10(N/m)
where N is the number of observations and m the number of series. Will be
automatically limited to one less than the number of observations in the series.}
}
\value{
A ggplot2 plot.
}
\description{
Plots a time series along with its ACF along with an customisable third
graphic of either a PACF, lagged scatterplot or spectral density.
}
\examples{
library(tsibble)
tsibbledata::ausretail \%>\%
  filter(
    State == "Victoria",
    Industry == "Cafes, restaurants and catering services"
  ) \%>\%
  ggtsdisplay(Turnover)

}
\references{
Hyndman and Athanasopoulos (2018) \emph{Forecasting: principles
and practice}, 2nd edition, OTexts: Melbourne, Australia.
\url{https://OTexts.org/fpp2/}
}
\seealso{
\code{\link[stats]{plot.ts}}, \code{\link{ACF}},
\code{\link[stats]{spec.ar}}
}
\author{
Rob J Hyndman & Mitchell O'Hara-Wild
}
