% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{hurst}
\alias{hurst}
\title{Hurst coefficient}
\usage{
hurst(x, ...)
}
\arguments{
\item{x}{a vector. If missing values are present, the largest
contiguous portion of the vector is used.}

\item{...}{Unused.}
}
\value{
A numeric value.
}
\description{
Computes the Hurst coefficient indicating the level of fractional differencing
of a time series.
}
\author{
Rob J Hyndman
}
