% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{as_tsibble}
\alias{autoplot}
\alias{autolayer}
\alias{components}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fablelite}{\code{\link[fablelite]{\%>\%}}, \code{\link[fablelite]{components}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}, \code{\link[ggplot2]{autolayer}}}

  \item{tsibble}{\code{\link[tsibble]{as_tsibble}}}
}}

