% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seats.R
\name{SEATS}
\alias{SEATS}
\title{Seasonal decomposition with X-13ARIMA-SEATS}
\usage{
SEATS(.data, formula, ...)
}
\arguments{
\item{.data}{A tsibble.}

\item{formula}{Decomposition specification.}

\item{...}{Other arguments passed to \code{\link[seasonal:seas]{seasonal::seas()}}.}
}
\description{
Decomposes a time series via model based signal extraction using X-13ARIMA-SEATS.
}
\examples{
tsibbledata::aus_production \%>\% SEATS(Beer)

}
\references{
Dagum, E. B., & Bianconcini, S. (2016) "Seasonal adjustment methods and real
time trend-cycle estimation". \emph{Springer}.

X-13ARIMA-SEATS Documentation from the seasonal package's website:
http://www.seasonal.website/seasonal.html

Official X-13ARIMA-SEATS manual: https://www.census.gov/ts/x13as/docX13ASHTML.pdf
}
\seealso{
\code{\link[seasonal:seas]{seasonal::seas()}}
}
