% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl.R
\name{STL}
\alias{STL}
\title{Multiple seasonal decomposition by Loess}
\usage{
STL(data, formula, iterations = 2, ...)
}
\arguments{
\item{data}{A tsibble.}

\item{formula}{Decomposition specification.}

\item{iterations}{Number of iterations to use to refine the seasonal component.}

\item{...}{Other arguments passed to \code{\link[forecast]{mstl}}.}
}
\description{
Multiple seasonal decomposition by Loess
}
\examples{
USAccDeaths \%>\% STL(value ~ season(window = 10))

}
\seealso{
\code{\link[stats]{stl}}, \code{\link[stats]{supsmu}}
}
